created header file for lcd interface
